Type warnings removed

+++ Colony/colony_t.cpp
@@ -45,7 +45,7 @@ void colony_t::print_info ( void )
     int i = 0;
 
     utilities_t::get()->print_header();
-    for( ; i < m_population.size(); i++ )
+    for( ; i < (int)m_population.size(); i++ )
     {
         utilities_t::get()->print_row( i + 1, m_population[i]->get_name(), m_population[i]->get_age(), m_population[i]->get_colour(),
                             m_population[i]->get_sex_string() );
@@ -59,7 +59,7 @@ void colony_t::grow ( void )
     rabbit_t *p_rabbit = nullptr;
     bool make_babies = contains_male();
     unsigned long max = m_population.size();
-    int i;
+    unsigned long i;
 
     for( i = 0; i < max; i++ )
     {
@@ -94,7 +94,7 @@ void colony_t::grow ( void )
 
 bool colony_t::has_food_shortage ( void ) const
 {
-    return m_population.size() >= m_FOOD_LIMIT;
+    return m_population.size() >= (unsigned long)m_FOOD_LIMIT;
 }
 
 bool colony_t::contains_male ( void ) const
@@ -116,7 +116,7 @@ bool colony_t::contains_male ( void ) const
 void colony_t::cull ( void )
 {
     unsigned long number_of_victims = m_population.size() / 2;
-    for( int i = 0 ; i < number_of_victims; i++ )
+    for( unsigned long i = 0 ; i < number_of_victims; i++ )
     {
         kill_wabbit( m_population[ utilities_t::get()->get_random_number_in_range( m_population.size() ) ] );
     }
